+++ CHANGELOG.md
@@ -30,6 +30,7 @@ to the project during `#hactoberfest`. List of awesome people:
 - Forbid `try` with `finally` without `except`
 - Forbit `if` statements with invalid conditionals
 - Forbid opening parenthesis from following keyword without space in between them
+- Forbid the use of more than 2 `for` loops within a comprehension
 
 ### Bugfixes